Combine lite binaries
