+++ .github/workflows/golangci-lint.yml
@@ -49,17 +49,11 @@ jobs:
         with:
           version: v2.12
 
-      - name: golangci-lint (cosign-sign)
+      - name: golangci-lint (cosign-lite)
         uses: golangci/golangci-lint-action@1e7e51e771db61008b38414a730f564565cf7c20 # v9.2.0
         with:
           version: v2.9
-          working-directory: cmd/cosign-sign
-
-      - name: golangci-lint (cosign-verify)
-        uses: golangci/golangci-lint-action@1e7e51e771db61008b38414a730f564565cf7c20 # v9.2.0
-        with:
-          version: v2.9
-          working-directory: cmd/cosign-verify
+          working-directory: cmd/cosign-lite
 
   golangci-test-e2e:
     name: lint-test-e2e
