+++ .goreleaser.yml
@@ -11,9 +11,8 @@ env:
 before:
   hooks:
     - go mod tidy
-    - go mod tidy -C cmd/cosign-sign
-    - go mod tidy -C cmd/cosign-verify
-    - git --no-pager diff --exit-code go.mod go.sum cmd/cosign-sign/go.mod cmd/cosign-sign/go.sum cmd/cosign-verify/go.mod cmd/cosign-verify/go.sum
+    - go mod tidy -C cmd/cosign-lite
+    - git --no-pager diff --exit-code go.mod go.sum cmd/cosign-lite/go.mod cmd/cosign-lite/go.sum
 
 gomod:
   proxy: true
@@ -153,10 +152,10 @@ builds:
       - pivkey
       - pkcs11key
 
-  - id: cosign-sign-linux
-    binary: cosign-sign-linux-{{ .Arch }}
+  - id: cosign-lite-linux
+    binary: cosign-lite-linux-{{ .Arch }}
     no_unique_dist_dir: true
-    dir: cmd/cosign-sign
+    dir: cmd/cosign-lite
     main: .
     flags:
       - -trimpath
@@ -175,51 +174,10 @@ builds:
     env:
       - CGO_ENABLED=0
 
-  - id: cosign-sign-darwin-windows
-    binary: cosign-sign-{{ .Os }}-{{ .Arch }}
+  - id: cosign-lite-darwin-windows
+    binary: cosign-lite-{{ .Os }}-{{ .Arch }}
     no_unique_dist_dir: true
-    dir: cmd/cosign-sign
-    main: .
-    flags:
-      - -trimpath
-    mod_timestamp: '{{ .CommitTimestamp }}'
-    goos:
-      - darwin
-      - windows
-    goarch:
-      - amd64
-      - arm64
-    ldflags:
-      - "{{ .Env.LDFLAGS }}"
-    env:
-      - CGO_ENABLED=0
-
-  - id: cosign-verify-linux
-    binary: cosign-verify-linux-{{ .Arch }}
-    no_unique_dist_dir: true
-    dir: cmd/cosign-verify
-    main: .
-    flags:
-      - -trimpath
-    mod_timestamp: '{{ .CommitTimestamp }}'
-    goos:
-      - linux
-    goarch:
-      - amd64
-      - arm64
-      - arm
-      - s390x
-      - ppc64le
-      - riscv64
-    ldflags:
-      - "{{ .Env.LDFLAGS }}"
-    env:
-      - CGO_ENABLED=0
-
-  - id: cosign-verify-darwin-windows
-    binary: cosign-verify-{{ .Os }}-{{ .Arch }}
-    no_unique_dist_dir: true
-    dir: cmd/cosign-verify
+    dir: cmd/cosign-lite
     main: .
     flags:
       - -trimpath
