+++ Makefile
@@ -81,11 +81,8 @@ log-%:
 cosign: $(SRCS)
 	CGO_ENABLED=0 $(GOEXE) build -trimpath -ldflags "$(LDFLAGS)" -o $@ ./cmd/cosign
 
-cosign-sign: $(SRCS)
-	CGO_ENABLED=0 $(GOEXE) build -C cmd/cosign-sign -trimpath -ldflags "$(LDFLAGS)" -o ../../$@ .
-
-cosign-verify: $(SRCS)
-	CGO_ENABLED=0 $(GOEXE) build -C cmd/cosign-verify -trimpath -ldflags "$(LDFLAGS)" -o ../../$@ .
+cosign-lite: $(SRCS)
+	CGO_ENABLED=0 $(GOEXE) build -C cmd/cosign-lite -trimpath -ldflags "$(LDFLAGS)" -o ../../$@ .
 
 cosign-pivkey-pkcs11key: $(SRCS)
 	CGO_ENABLED=1 $(GOEXE) build -trimpath -tags=pivkey,pkcs11key -ldflags "$(LDFLAGS)" -o cosign ./cmd/cosign
@@ -120,8 +117,7 @@ test:
 
 clean:
 	rm -rf cosign
-	rm -rf cosign-sign
-	rm -rf cosign-verify
+	rm -rf cosign-lite
 	rm -rf dist/
 
 KOCACHE_PATH=/tmp/ko
@@ -231,6 +227,5 @@ include test/ci.mk
 .PHONY: docgen
 docgen:
 	$(GOEXE) run -tags pivkey,pkcs11key,cgo ./cmd/help/
-	cd cmd/cosign-sign && $(GOEXE) run ./help --dir ../../doc
-	cd cmd/cosign-verify && $(GOEXE) run ./help --dir ../../doc
+	cd cmd/cosign-lite && $(GOEXE) run ./help --dir ../../doc
