+++ cmd/conformance/main.go
@@ -90,8 +90,7 @@ func main() {
 
 	args := []string{}
 
-	useCosignSign := env.Getenv("COSIGN_SIGN_CONFORMANCE") == "true" && os.Args[1] == "sign-bundle"
-	useCosignVerify := env.Getenv("COSIGN_VERIFY_CONFORMANCE") == "true" && os.Args[1] == "verify-bundle"
+	useCosignLite := env.Getenv("COSIGN_LITE_CONFORMANCE") == "true"
 
 	switch os.Args[1] {
 	case "sign-bundle":
@@ -99,14 +98,14 @@ func main() {
 		if inToto {
 			cmdName = "attest"
 		}
-		if !useCosignSign {
+		if !useCosignLite {
 			cmdName += "-blob"
 		}
 		args = append(args, cmdName)
 		args = append(args, "-y")
 
 	case "verify-bundle":
-		if useCosignVerify {
+		if useCosignLite {
 			if inToto {
 				args = append(args, "verify-attestation")
 			} else {
@@ -138,7 +137,7 @@ func main() {
 
 	if bundlePath != nil {
 		args = append(args, "--bundle", *bundlePath)
-		if !useCosignSign && !useCosignVerify {
+		if !useCosignLite {
 			args = append(args, "--new-bundle-format")
 		}
 	}
@@ -160,18 +159,15 @@ func main() {
 	if keyPath != nil {
 		args = append(args, "--key", *keyPath)
 	}
-	if inToto && !useCosignSign && !useCosignVerify {
+	if inToto && !useCosignLite {
 		args = append(args, "--statement")
 	}
 	args = append(args, os.Args[len(os.Args)-1])
 
 	dir := filepath.Dir(os.Args[0])
 	binaryName := "cosign"
-	switch {
-	case useCosignSign:
-		binaryName = "cosign-sign"
-	case useCosignVerify:
-		binaryName = "cosign-verify"
+	if useCosignLite {
+		binaryName = "cosign-lite"
 	}
 
 	initCmd := exec.Command(filepath.Join(dir, binaryName), "initialize") // #nosec G204,G702 -- conformance harness invokes the sibling cosign binary
