+++ cmd/cosign-lite/cli/initialize.go
