+++ cmd/cosign-lite/cli/initialize_test.go
