+++ cmd/cosign-lite/cli/keys.go
