+++ cmd/cosign-lite/cli/keys_test.go
