+++ cmd/cosign-lite/cli/sign.go
