+++ cmd/cosign-lite/cli/sign_options.go
