+++ cmd/cosign-lite/cli/sign_test.go
