--- cmd/cosign-sign/main.go
@@ -1,31 +1,0 @@
-// Copyright 2026 The Sigstore Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"os"
-
-	"github.com/sigstore/cosign/v3/cmd/cosign-sign/cli"
-
-	_ "github.com/sigstore/cosign/v3/pkg/providers/envvar"
-	_ "github.com/sigstore/cosign/v3/pkg/providers/filesystem"
-	_ "github.com/sigstore/cosign/v3/pkg/providers/github"
-)
-
-func main() {
-	if err := cli.New().Execute(); err != nil {
-		os.Exit(1)
-	}
-}
