+++ cmd/help/verify.sh
@@ -19,7 +19,7 @@ set -e
 # Verify that generated Markdown docs are up-to-date.
 tmpdir=$(mktemp -d)
 go run -tags pivkey,pkcs11key,cgo cmd/help/main.go --dir "$tmpdir"
-(cd cmd/cosign-sign && go run ./help --dir "$tmpdir")
+(cd cmd/cosign-lite && go run ./help --dir "$tmpdir")
 echo "###########################################"
 echo "If diffs are found, run: make docgen"
 echo "###########################################"
