+++ doc/cosign-lite.md
@@ -1,0 +1,19 @@
+## cosign-lite
+
+cosign-lite is a lightweight Sigstore signing and verification utility
+
+### Options
+
+```
+  -h, --help   help for cosign-lite
+```
+
+### SEE ALSO
+
+* [cosign-lite attest](cosign-lite_attest.md)	 - Sign the supplied in-toto statement and generate a Sigstore bundle with a DSSE envelope
+* [cosign-lite generate-key-pair](cosign-lite_generate-key-pair.md)	 - Generate a local password-encrypted key pair
+* [cosign-lite initialize](cosign-lite_initialize.md)	 - Initialize TUF roots of trust
+* [cosign-lite sign](cosign-lite_sign.md)	 - Sign the supplied payload blob and generate a Sigstore bundle
+* [cosign-lite verify](cosign-lite_verify.md)	 - Verify a plain payload blob against a standardized Sigstore bundle
+* [cosign-lite verify-attestation](cosign-lite_verify-attestation.md)	 - Verify a signed in-toto statement and output its payload to stdout
+
