--- doc/cosign-sign.md
@@ -1,17 +1,0 @@
-## cosign-sign
-
-cosign-sign is a minimal signing utility for Sigstore
-
-### Options
-
-```
-  -h, --help   help for cosign-sign
-```
-
-### SEE ALSO
-
-* [cosign-sign attest](cosign-sign_attest.md)	 - Sign the supplied in-toto statement and generate a Sigstore bundle with a DSSE envelope
-* [cosign-sign generate-key-pair](cosign-sign_generate-key-pair.md)	 - Generate a local password-encrypted key pair
-* [cosign-sign initialize](cosign-sign_initialize.md)	 - Initialize TUF roots of trust
-* [cosign-sign sign](cosign-sign_sign.md)	 - Sign the supplied payload blob and generate a Sigstore bundle
-
