--- doc/cosign-verify.md
@@ -1,16 +1,0 @@
-## cosign-verify
-
-cosign-verify is a minimal verification utility for Sigstore
-
-### Options
-
-```
-  -h, --help   help for cosign-verify
-```
-
-### SEE ALSO
-
-* [cosign-verify initialize](cosign-verify_initialize.md)	 - Initialize TUF roots of trust
-* [cosign-verify verify](cosign-verify_verify.md)	 - Verify a plain payload blob against a standardized Sigstore bundle
-* [cosign-verify verify-attestation](cosign-verify_verify-attestation.md)	 - Verify a signed in-toto statement and output its payload to stdout
-
